Fix widget view file split
